feat: show other sectors
